refactor: move FileDetailDto.class from file.dto.request to file.dto
